+++ src/styles/fonts.ts
@@ -1,0 +1,5 @@
+const AdamWarren = require("../fonts/adam.ttf")
+
+export {
+  AdamWarren
+};